Reorganize the components of the data module
