+++ packages/data/src/components/with-select/index.js
@@ -11,8 +11,8 @@ import isShallowEqual from '@wordpress/is-shallow-equal';
 /**
  * Internal dependencies
  */
-import defaultRegistry from '../default-registry';
-import { withRegistry } from './registry-provider';
+import defaultRegistry from '../../default-registry';
+import { withRegistry } from '../registry-provider';
 
 /**
  * Higher-order component used to inject state-derived props using registered
